refactor: SetAdditionalInfo.tsx component refactoring
